+++ src/http/mod.rs
@@ -2,4 +2,4 @@ mod client;
 mod request;
 
 pub use client::create_client;
-pub use request::execute_request;
+pub use request::{execute_request, now_us};